챗봇 test용 코드들 삭제 및 채팅 규칙 추가
